git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@696 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ fs.h
@@ -29,8 +29,8 @@ typedef struct FsPublic_t {
 
 	int serialized;
 	unsigned long serial_number;
-	unsigned int cluster_size;
-	unsigned int sector_size;
+	uint8_t cluster_size;
+	uint16_t sector_size;
 } FsPublic_t;
 
 Stream_t *fs_init(char drive, int mode, int *isRop);
